Fix filename when no filenamePrefix is defined
When generating the output filename, if no filenamePrefix is defined the
file name should not include any prefix.
The filename generation was concatenating the prefix in the path
section, resulting in a filename starting with `-` when no
filenamePrefix is specified.

The fix consists of splitting the path and filename generation.
The filename generation is responsible of prepending the filenamePrefix
and then the path and filename are concatenated.