Add sync-public-main GH action script (#6)
This script runs only on the private repo and it's needed to synchronize
master branch of the private repository with the main branch of the public
repository. This is done updating the public-main branch and manually
triggering this script in GitHub.

It also updates existing workflows to run only on private repo.